+++ src/main/java/com/ciphertool/engine/fitness/cipherkey/ConstrainedTieredSamplingMarkovFitnessEvaluator.java
@@ -19,6 +19,7 @@
 
 package com.ciphertool.engine.fitness.cipherkey;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -86,7 +87,7 @@ public class ConstrainedTieredSamplingMarkovFitnessEvaluator implements FitnessE
 	}
 
 	@Override
-	public Double evaluate(Chromosome chromosome) {
+	public BigDecimal evaluate(Chromosome chromosome) {
 		CipherKeyChromosome cipherKeyChromosome = (CipherKeyChromosome) chromosome;
 
 		Map<Character, Integer> actualLetterCounts = new HashMap<Character, Integer>(LOWERCASE_LETTERS.size());
@@ -166,7 +167,7 @@ public class ConstrainedTieredSamplingMarkovFitnessEvaluator implements FitnessE
 			nGramProbability = 0.0;
 		}
 
-		return (nGramProbability * nGramWeight) + (frequencyProbability * frequencyWeight);
+		return BigDecimal.valueOf((nGramProbability * nGramWeight) + (frequencyProbability * frequencyWeight));
 	}
 
 	@Override
