+++ src/main/java/com/ciphertool/engine/fitness/cipherkey/MatchingWordGraphCorpusFitnessEvaluator.java
@@ -16,6 +16,7 @@
 
 package com.ciphertool.engine.fitness.cipherkey;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -268,7 +269,7 @@ public class MatchingWordGraphCorpusFitnessEvaluator implements FitnessEvaluator
 	}
 
 	@Override
-	public Double evaluate(Chromosome chromosome) {
+	public BigDecimal evaluate(Chromosome chromosome) {
 		String currentSolutionString = WordGraphUtils.getSolutionAsString((CipherKeyChromosome) chromosome).substring(0, lastRowBegin);
 
 		Map<Integer, Match> matchMap = new HashMap<Integer, Match>(currentSolutionString.length());
@@ -348,7 +349,7 @@ public class MatchingWordGraphCorpusFitnessEvaluator implements FitnessEvaluator
 			log.debug("Best branch: " + bestBranch);
 		}
 
-		return highestScore;
+		return BigDecimal.valueOf(highestScore);
 	}
 
 	private static List<Integer> getMatchIndices(int lastRowBegin, String bestBranch, String currentSolutionString) {
