+++ src/main/java/com/ciphertool/engine/fitness/cipherkey/WordGraphFitnessEvaluator.java
@@ -19,6 +19,7 @@
 
 package com.ciphertool.engine.fitness.cipherkey;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -43,7 +44,7 @@ public class WordGraphFitnessEvaluator implements FitnessEvaluator {
 	private int			lastRowBegin;
 
 	@Override
-	public Double evaluate(Chromosome chromosome) {
+	public BigDecimal evaluate(Chromosome chromosome) {
 		String currentSolutionString = WordGraphUtils.getSolutionAsString((CipherKeyChromosome) chromosome).substring(0, lastRowBegin);
 
 		Map<Integer, Match> matchMap = new HashMap<Integer, Match>();
@@ -101,7 +102,7 @@ public class WordGraphFitnessEvaluator implements FitnessEvaluator {
 			}
 		}
 
-		return highestScore;
+		return BigDecimal.valueOf(highestScore);
 	}
 
 	@Override
